Initial poking around at how to make threads in java
We don't really understand how to create and call threads yet, or what's going on with the start method that's called
from main, so this commit is mostly us just flailing around a bit.

Based off of our reading, it seems like we need to make some class that implements runnable, that then contains a run
method that performs the usual reading or writing activity we would normally do.